Remove mini assembler code for byte swapping
The byte swapping was realized using inline assembler code using the bswap
instruction. This was less portable and even resulted in build failures when
compiling for OSX. It also reduced the effectivness of the optimizer slightly.

Instead the builtin functions of the compiler can be used to do the same.